refactor: Add setup for every api group
